+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 evaluation. This is useful e.g. in afterburner calculations.
 
 ### Changed
+* Minimum supported CMake version is now version 3.23
 * The `make install` target has been standardized to create an enclosing version-specific folder inside which the standard `bin`, `lib`, `include` and `share` directories are created (previously, the version specific folder was created inside the standard ones).
 
 ## SMASH-3.3